+++ doc/tutorials/QuartusTutorial/Files/pinout.tcl
@@ -1,0 +1,19 @@
+##Clock signal
+set_location_assignment PIN_M9 -to Clk
+
+##Switches
+set_location_assignment PIN_U13 -to Switches[0]
+set_location_assignment PIN_V13 -to Switches[1]
+set_location_assignment PIN_T13 -to Switches[2]
+set_location_assignment PIN_T12 -to Switches[3]
+
+##Buttons
+set_location_assignment PIN_W9 -to Buttons[1]
+set_location_assignment PIN_U7 -to Buttons[0]
+
+##LEDs
+set_location_assignment PIN_AA2 -to Led[0]
+set_location_assignment PIN_AA1 -to Led[1]
+set_location_assignment PIN_W2 -to Led[2]
+set_location_assignment PIN_Y3 -to Led[3]
+